+++ src/Strategos.Ontology/Descriptors/RequiredEdgeProperty.cs
@@ -1,8 +1,0 @@
-namespace Strategos.Ontology.Descriptors;
-
-public sealed record RequiredEdgeProperty
-{
-    public required string Name { get; init; }
-
-    public required string TypeName { get; init; }
-}